Arreglos metodo edit

+++ src/main/java/nicoangeletti/bazar/dto/ClaseDTO.java
@@ -12,12 +12,12 @@ public class ClaseDTO {
     
     
     private Long codigo_venta;
-    private int total;
+    private Double total;
     private int cantidad_productos;
     private String nombre_cliente;
     private String apellido_cliente;
 
-    public ClaseDTO(Long codigo_venta, int total, int cantidad_productos, String nombre_cliente, String apellido_cliente) {
+    public ClaseDTO(Long codigo_venta, Double total, int cantidad_productos, String nombre_cliente, String apellido_cliente) {
         this.codigo_venta = codigo_venta;
         this.total = total;
         this.cantidad_productos = cantidad_productos;
